Test case firing issue 12712
The scenario is that we have two identical files defining test cases and suites. The suites are the same in both files, tests are not. The line numbers are the same as well.
Finally those files are #include'd in a unique .cpp file. The registrar for the test suites are having the same name and raising a compilation error